Move a bunch of totally non-jiggy code out of e2eservice.TestJig
When test/e2e/framework/service was first created, a lot of service
test code got crammed into TestJig even though it didn't use any of
TestJig's members. Make that code available outside of TestJig.
